# server/media/docs/swagger.yaml
components: {}
externalDocs:
  description: ""
  url: ""
info:
  description: Микросервис для работы с медиа-файлами в Bloom
  title: Bloom Media API
  version: "1.0"
openapi: 3.1.0
paths:
  /media/{id}:
    get:
      description: Отдает содержимое файла по его ID
      parameters:
      - description: ID медиа-файла
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: string
          description: Содержимое файла
        "400":
          content:
            application/octet-stream:
              schema:
                type: string
          description: ID is required
        "404":
          content:
            application/octet-stream:
              schema:
                type: string
          description: File not found
        "500":
          content:
            application/octet-stream:
              schema:
                type: string
          description: Internal server error
      summary: Скачать/Просмотреть медиа-файл
      tags:
      - media
  /media/upload:
    post:
      description: Загружает файл на сервер и делает его временным до подтверждения
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  description: Файл для загрузки
                  format: binary
                  type: string
              required:
              - file
              type: object
        description: Файл для загрузки
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: 'Пример: {\"url\": \"/api/v1/media/uuid\"}'
        "400":
          content:
            application/json:
              schema:
                type: string
          description: Invalid file
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Internal server error
      summary: Загрузить медиа-файл
      tags:
      - media